Fix: 1. data type issues 2. recovery feedrate and flowrate at recovery_env function

+++ snapmaker/src/module/toolhead_fdm.h
@@ -117,8 +117,8 @@ typedef enum {
 
 typedef struct {
   uint8_t active_extruder;
-  float feedrate_percentage[EXTRUDERS];
-  float flowrate_percentage[EXTRUDERS];
+  int16_t feedrate_percentage[EXTRUDERS];
+  int16_t flowrate_percentage[EXTRUDERS];
   float live_z_offset[EXTRUDERS];
   float live_z_offset_changed;
   uint8_t fan_speed[3];
